+++ src/lib/FormCheckbox.php
@@ -9,8 +9,10 @@ class FormCheckbox extends FormInput {
 		$this->name = $name;
 		$this->label = $label;
 		$this->hint = null;
+		$this->text = '';
 
 		$this->pop_attr('hint', $attr, $this->hint);
+		$this->pop_attr('text', $attr, $this->text);
 
 		if (  $type != null  ) $attr['type'] = $type;
 		if (    $id != null  ) $attr['id'] = $id;
@@ -24,12 +26,14 @@ class FormCheckbox extends FormInput {
 		$this->attr = $attr;
 	}
 
+	public function get_text(){
+		return $this->text;
+	}
 
 	public function get_content(array $extra_attr = array(), array $label = array()){
 		$attr = array_merge_recursive($extra_attr, $this->attr);
-		$text = $this->label;
 		if ( $this->get_container() instanceof FormGroup ){
-			return "<label " . FormUtils::serialize_attr($label) . "><input " . $this->serialize_attr($attr) . " />$text</label>";
+			return "<label " . FormUtils::serialize_attr($label) . "><input " . $this->serialize_attr($attr) . " />{$this->text}</label>";
 		} else {
 			return "<input " . $this->serialize_attr($attr) . " />";
 		}